Refactor raster clipping script
