+++ geospatial_tools/raster.py
@@ -1,11 +1,19 @@
 """This module contains functions that process or create raster/image data."""
 
+import concurrent.futures
+import logging
 import pathlib
+from concurrent.futures import ProcessPoolExecutor
+from multiprocessing import cpu_count
 from typing import Union
 
+import geopandas as gpd
 import rasterio
+from geopandas import GeoDataFrame
+from rasterio.mask import mask
 from rasterio.warp import Resampling, calculate_default_transform, reproject
 
+from geospatial_tools import DATA_DIR
 from geospatial_tools.utils import create_crs, create_logger
 
 LOGGER = create_logger(__name__)
@@ -15,7 +23,7 @@ def reproject_raster(
     dataset_path: Union[str, pathlib.Path],
     target_crs: Union[str, int],
     target_path: Union[str, pathlib.Path],
-    logger=LOGGER,
+    logger: logging.Logger = LOGGER,
 ):
     """
 
@@ -60,3 +68,61 @@ def reproject_raster(
     if target_path.exists():
         logger.info(f"Reprojected file created at {target_path}")
         return target_path
+
+
+def clip_process(
+    raster_image: Union[pathlib.Path, str],
+    id_polygon: tuple[int, GeoDataFrame],
+    s2_tile_id: str,
+    output_path: Union[pathlib.Path, str],
+):
+    try:
+        polygon_id, polygon = id_polygon
+        with rasterio.open(raster_image) as src:
+            out_image, out_transform = mask(src, [polygon], crop=True)
+            out_meta = src.meta.copy()
+        out_meta.update(
+            {"driver": "GTiff", "height": out_image.shape[1], "width": out_image.shape[2], "transform": out_transform}
+        )
+        if isinstance(output_path, str):
+            output_path = pathlib.Path(output_path)
+        output_file = output_path / f"{s2_tile_id}_clipped_{polygon_id}.tif"
+        with rasterio.open(output_file, "w", **out_meta) as dest:
+            dest.write(out_image)
+
+        return polygon_id, polygon
+    except Exception as e:  # pylint: disable=broad-exception-caught
+        return str(e)
+
+
+def clip_raster_with_polygon(
+    raster_image: Union[pathlib.Path, str],
+    polygon_layer: Union[pathlib.Path, str, GeoDataFrame],
+    s2_tile_id: str,
+    output_path: Union[str, pathlib.Path] = DATA_DIR,
+    num_processes: int = None,
+    logger: logging.Logger = LOGGER,
+):
+    workers = cpu_count()
+    if num_processes:
+        workers = num_processes
+
+    logger.info(f"Number of workers used: {workers}")
+    logger.info(f"Output path : [{output_path}]")
+
+    gdf = polygon_layer
+    if not isinstance(polygon_layer, GeoDataFrame):
+        gdf = gpd.read_file(polygon_layer)
+
+    polygons = gdf["geometry"]
+    ids = gdf.index
+
+    id_polygon_list = zip(ids, polygons)
+    with ProcessPoolExecutor(max_workers=workers) as executor:
+        futures = [
+            executor.submit(clip_process, raster_image, polygon_tuple, s2_tile_id, output_path)
+            for polygon_tuple in id_polygon_list
+        ]
+        results = [future.result() for future in concurrent.futures.as_completed(futures)]
+    for result in results:
+        logger.info(f"Writing file successful : [{result}]")
